componentize filter bar and tabs
